add nation to post and permission to employee

+++ source/src/main/java/com/nextstep/api/form/nation/CreateNationForm.java
@@ -22,7 +22,7 @@ public class CreateNationForm{
     private Integer kind;
     @NotNull(message = "status cant not be null")
     @ApiModelProperty(name = "status", required = true)
-    private Long status;
+    private int status;
     @ApiModelProperty(name = "parentId")
     private Long parentId;
 }
